% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelist_wfs.R
\name{thelist_layer}
\alias{thelist_layer}
\title{Find layer by pattern}
\usage{
thelist_layer(x, random = FALSE)
}
\arguments{
\item{x}{string pattern to match, e.g. "cadastral"}

\item{random}{allow random selection of one ignoring 'x', 'FALSE' by default}
}
\value{
character string, one of \link{list_layers}
}
\description{
List layers have names like "Public_OpenDataWFS:LIST_Cadastral_Parcels", so we have a
simple string lookup to find the right layer.
}
\details{
Approximate matching is used, if you need to be exact look at \link{list_layers} and choose it deliberately.
}
\examples{
thelist_layer("cadastral")
thelist_layer("cadastre")
thelist_layer(random = TRUE)
}
