% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thelist_wfs.R
\name{thelist_wfs}
\alias{thelist_wfs}
\title{TheLIST WFS service}
\usage{
thelist_wfs()
}
\value{
data source name of TheLIST WFS Server
}
\description{
TheLIST WFS service
}
\examples{
thelist_wfs()
library(lazysf)
address <- lazysf(thelist_wfs(), thelist_layer("address"))
address
}
